Remove old generator.
